# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Fill-in-the-blank Quiz System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Question {
  id: number;
  /** Optional label for the question */
  title: string;
  /** The full question text with {{answer}} placeholders marking blank spots */
  text: string;
  /** The correct answers in order of appearance */
  answers: string[];
  totalAttempts: number;
  correctAttempts: number;
  hasWrongAttempt: boolean;
  createdAt: string;
}

export interface CreateQuestionBody {
  title: string;
  /** Question text with {{answer}} placeholders */
  text: string;
}

export interface SubmitAttemptBody {
  /** User's answers in order */
  answers: string[];
}

export interface AttemptResult {
  correct: boolean;
  correctAnswers: string[];
  userAnswers: string[];
  perAnswerCorrect: boolean[];
}

export interface QuizStats {
  totalQuestions: number;
  questionsWithWrongAttempts: number;
  totalAttempts: number;
  correctAttempts: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListQuestionsParams = {
  /**
   * If true, return only questions that have been answered incorrectly at least once
   */
  wrongOnly?: boolean;
};