# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Fill-in-the-blank Quiz System
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all questions
 */
export const ListQuestionsQueryParams = zod.object({
  wrongOnly: zod.coerce
    .boolean()
    .optional()
    .describe(
      "If true, return only questions that have been answered incorrectly at least once",
    ),
});

export const ListQuestionsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string().describe("Optional label for the question"),
  text: zod
    .string()
    .describe(
      "The full question text with {{answer}} placeholders marking blank spots",
    ),
  answers: zod
    .array(zod.string())
    .describe("The correct answers in order of appearance"),
  totalAttempts: zod.number(),
  correctAttempts: zod.number(),
  hasWrongAttempt: zod.boolean(),
  createdAt: zod.date(),
});
export const ListQuestionsResponse = zod.array(ListQuestionsResponseItem);

/**
 * @summary Create a new question
 */
export const CreateQuestionBody = zod.object({
  title: zod.string(),
  text: zod.string().describe("Question text with {{answer}} placeholders"),
});

/**
 * @summary Get a single question
 */
export const GetQuestionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetQuestionResponse = zod.object({
  id: zod.number(),
  title: zod.string().describe("Optional label for the question"),
  text: zod
    .string()
    .describe(
      "The full question text with {{answer}} placeholders marking blank spots",
    ),
  answers: zod
    .array(zod.string())
    .describe("The correct answers in order of appearance"),
  totalAttempts: zod.number(),
  correctAttempts: zod.number(),
  hasWrongAttempt: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Update a question
 */
export const UpdateQuestionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateQuestionBody = zod.object({
  title: zod.string(),
  text: zod.string().describe("Question text with {{answer}} placeholders"),
});

export const UpdateQuestionResponse = zod.object({
  id: zod.number(),
  title: zod.string().describe("Optional label for the question"),
  text: zod
    .string()
    .describe(
      "The full question text with {{answer}} placeholders marking blank spots",
    ),
  answers: zod
    .array(zod.string())
    .describe("The correct answers in order of appearance"),
  totalAttempts: zod.number(),
  correctAttempts: zod.number(),
  hasWrongAttempt: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a question
 */
export const DeleteQuestionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Submit an attempt for a question
 */
export const SubmitAttemptParams = zod.object({
  id: zod.coerce.number(),
});

export const SubmitAttemptBody = zod.object({
  answers: zod.array(zod.string()).describe("User's answers in order"),
});

export const SubmitAttemptResponse = zod.object({
  correct: zod.boolean(),
  correctAnswers: zod.array(zod.string()),
  userAnswers: zod.array(zod.string()),
  perAnswerCorrect: zod.array(zod.boolean()),
});

/**
 * @summary Reset attempt stats for a question
 */
export const ResetQuestionStatsParams = zod.object({
  id: zod.coerce.number(),
});

export const ResetQuestionStatsResponse = zod.object({
  id: zod.number(),
  title: zod.string().describe("Optional label for the question"),
  text: zod
    .string()
    .describe(
      "The full question text with {{answer}} placeholders marking blank spots",
    ),
  answers: zod
    .array(zod.string())
    .describe("The correct answers in order of appearance"),
  totalAttempts: zod.number(),
  correctAttempts: zod.number(),
  hasWrongAttempt: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Get overall quiz statistics
 */
export const GetStatsResponse = zod.object({
  totalQuestions: zod.number(),
  questionsWithWrongAttempts: zod.number(),
  totalAttempts: zod.number(),
  correctAttempts: zod.number(),
});