# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Fill-in-the-blank Quiz System
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  /** Optional label for the question */
  title: string;
  /** The full question text with {{answer}} placeholders marking blank spots */
  text: string;
  /** The correct answers in order of appearance */
  answers: string[];
  totalAttempts: number;
  correctAttempts: number;
  hasWrongAttempt: boolean;
  createdAt: Date;
}