# lib/api-zod/src/generated/types/quizStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Fill-in-the-blank Quiz System
 * OpenAPI spec version: 0.1.0
 */

export interface QuizStats {
  totalQuestions: number;
  questionsWithWrongAttempts: number;
  totalAttempts: number;
  correctAttempts: number;
}